+++ lang/et/charon.php
@@ -206,3 +206,9 @@ $string['availability'] = 'Kättesaadavus';
 $string['lab_teachers'] = 'Praktikumi õpetajad';
 $string['lab_queue'] = 'Kaitsmise ootejärjekord';
 $string['lab_end'] = 'Praktikum on lõppenud';
+
+/*------------------------*\
+   Messages
+\*------------------------*/
+
+$string['messageprovider:comment'] = 'Teavitus kommentaarist charoni katsele';